www: extract about component
